refactor(index.ts): move main logic to separate function and export it feat(local.ts): add local script for running GitHub action script locally docs(README.md): add section on running GitHub action script locally
